Extract config file
